working like stuff

+++ server/lib/final_project_web/controllers/like_controller.ex
@@ -3,6 +3,7 @@ defmodule FinalProjectWeb.LikeController do
 
   alias FinalProject.Likes
   alias FinalProject.Likes.Like
+  alias FinalProjectWeb.SessionController
 
   action_fallback FinalProjectWeb.FallbackController
 
@@ -11,12 +12,21 @@ defmodule FinalProjectWeb.LikeController do
     render(conn, "index.json", likes: likes)
   end
 
-  def create(conn, %{"like" => like_params}) do
-    with {:ok, %Like{} = like} <- Likes.create_like(like_params) do
+  def create(conn, %{"like" => like_params, "session" => session}) do
+    if SessionController.authorized?(conn, session["user_id"], session["token"]) do
+      with {:ok, %Like{} = _like} <- Likes.create_like(like_params) do
+        posts = FinalProject.Posts.get_recent_posts()
+        FinalProjectWeb.Endpoint.broadcast("feed:", "feed/update", FinalProjectWeb.PostView.render("index.json", posts: posts))
+
+        conn
+        |> send_resp(:created, "Success")
+      end
+    else
       conn
-      |> put_status(:created)
-      |> put_resp_header("location", Routes.like_path(conn, :show, like))
-      |> render("show.json", like: like)
+      |> put_resp_header(
+        "content-type",
+        "application/json; charset=UTF-8")
+      |> send_resp(:unauthorized, Jason.encode!(%{error: "Unauthorized"}))
     end
   end
 
@@ -33,11 +43,29 @@ defmodule FinalProjectWeb.LikeController do
     end
   end
 
-  def delete(conn, %{"id" => id}) do
+  def delete(conn, %{"id" => id, "session" => session}) do
     like = Likes.get_like!(id)
-
-    with {:ok, %Like{}} <- Likes.delete_like(like) do
-      send_resp(conn, :no_content, "")
+    user_id = session["user_id"]
+    if SessionController.authorized?(conn, user_id, session["token"]) do
+      if user_id == like.user_id do
+        with {:ok, %Like{}} <- Likes.delete_like(like) do
+          posts = FinalProject.Posts.get_recent_posts()
+          FinalProjectWeb.Endpoint.broadcast("feed:", "feed/update", FinalProjectWeb.PostView.render("index.json", posts: posts))
+          send_resp(conn, :no_content, "")
+        end
+      else
+        conn
+        |> put_resp_header(
+          "content-type",
+          "application/json; charset=UTF-8")
+        |> send_resp(:unauthorized, Jason.encode!(%{error: "Unauthorized"}))
+      end
+    else
+      conn
+      |> put_resp_header(
+        "content-type",
+        "application/json; charset=UTF-8")
+      |> send_resp(:unauthorized, Jason.encode!(%{error: "Unauthorized"}))
     end
   end
 end
